pyproject.toml
[project]
name = "bull-vision-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Hieu Tran",email = "hieutrantrung.it@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<=3.13.3"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "python-telegram-bot (>=20.8,<21.0)",
    "uvicorn (>=0.27.1,<0.28.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pydantic (>=2.6.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov=bot --cov-report=term-missing"
